Add tools to import Socrata traffic fatality data (#25)
This patch adds tools to download Socarata traffic fatality data sets
and an importer to include the content into a scrapd data set.

The data sets from the Socarata database were added into the
`external-datasets` directory.

The ScrAPD data sets for the years 2017 and 2018 were augemented with
the Socrata data. However, as the merge adds some entries that are
incomplete the new augmented data sets were not used for anything. The
may be merged into the final results once we are 100% convinced of the
data.